+++ Raptor/Factories/ContentLoaderFactory.cs
@@ -5,7 +5,13 @@
 namespace Raptor.Factories
 {
     using System.Diagnostics.CodeAnalysis;
+    using System.IO.Abstractions;
+    using Raptor.Audio;
     using Raptor.Content;
+    using Raptor.Graphics;
+    using Raptor.OpenAL;
+    using Raptor.OpenGL;
+    using Raptor.Services;
 
     /// <summary>
     /// Creates instances of a content loader.
@@ -13,10 +19,79 @@ namespace Raptor.Factories
     [ExcludeFromCodeCoverage]
     public static class ContentLoaderFactory
     {
+        private static IContentLoader? contentLoader;
+        private static ILoader<ITexture>? textureLoader;
+        private static ILoader<IAtlasData>? atlasLoader;
+        private static ILoader<ISound>? soundLoader;
+
         /// <summary>
         /// Creates a single instance of a content loader.
         /// </summary>
         /// <returns>A framework content loader implementation.</returns>
-        public static IContentLoader CreateContentLoader() => IoC.Container.GetInstance<IContentLoader>();
+        public static IContentLoader CreateContentLoader()
+        {
+            if (contentLoader is null)
+            {
+                contentLoader = new ContentLoader(CreateTextureLoader(), CreateSoundLoader(), CreateTextureAtlasLoader());
+            }
+
+            return contentLoader;
+        }
+
+        /// <summary>
+        /// Creates a loader that loads textures from disk.
+        /// </summary>
+        /// <returns>A loader for loading textures.</returns>
+        public static ILoader<ITexture> CreateTextureLoader()
+        {
+            if (textureLoader is null)
+            {
+                var glInvoker = IoC.Container.GetInstance<IGLInvoker>();
+                var imageService = IoC.Container.GetInstance<IImageFileService>();
+                IPathResolver texturePathResolver = new TexturePathResolver(IoC.Container.GetInstance<IDirectory>());
+
+                textureLoader = new TextureLoader(glInvoker, imageService, texturePathResolver);
+            }
+
+            return textureLoader;
+        }
+
+        /// <summary>
+        /// Creates a loader for loading atlas data from disk.
+        /// </summary>
+        /// <returns>A loader for loading texture atlas data.</returns>
+        public static ILoader<IAtlasData> CreateTextureAtlasLoader()
+        {
+            if (atlasLoader is null)
+            {
+                var textureLoader = new TextureLoader(IoC.Container.GetInstance<IImageFileService>(), IoC.Container.GetInstance<AtlasTexturePathResolver>());
+
+                IPathResolver atlasDataPathResolver = new AtlasJSONDataPathResolver(IoC.Container.GetInstance<IDirectory>());
+
+                atlasLoader = new AtlasLoader(textureLoader, atlasDataPathResolver, IoC.Container.GetInstance<IFile>());
+            }
+
+            return atlasLoader;
+        }
+
+        /// <summary>
+        /// Creates a loader that loads sounds from disk.
+        /// </summary>
+        /// <returns>A loader for loading sound data.</returns>
+        public static ILoader<ISound> CreateSoundLoader()
+        {
+            if (soundLoader is null)
+            {
+                var alInvoker = IoC.Container.GetInstance<IALInvoker>();
+                var audioManager = AudioDeviceManagerFactory.CreateDeviceManager();
+                IPathResolver soundPathResolver = new SoundPathResolver(IoC.Container.GetInstance<IDirectory>());
+                var oggDecoder = IoC.Container.GetInstance<ISoundDecoder<float>>();
+                var mp3Decoder = IoC.Container.GetInstance<ISoundDecoder<byte>>();
+
+                soundLoader = new SoundLoader(alInvoker, audioManager, soundPathResolver, oggDecoder, mp3Decoder);
+            }
+
+            return soundLoader;
+        }
     }
 }